Keep forwarding errors with extended ACKs
Extended ACKs are a common way for the kernel to indicate warnings
by not setting an error code, but providing more information about
the warning in the extended ACK.

An example for that is
https://github.com/torvalds/linux/blob/5bd7ef53ffe5ca580e93e74eb8c81ed191ddc4bd/net/sched/sch_taprio.c#L2004
And iproute2 prints this warning accordingly
https://github.com/iproute2/iproute2/blob/1a68525f4613b4e02e83d4b8004f22ac7ecbfedf/lib/libnetlink.c#L73

Therefore, forward NLMSG_ERROR message without error code
but extended ACKs to the user application
and let it decide how to handle it (e.g. print, ignore or fail).

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>